+++ src/Api/firebase/config.ts
@@ -1,4 +1,4 @@
-const firebaseConfig = {
+const config = {
   apiKey: process.env.REACT_APP_API_KEY,
   authDomain: process.env.REACT_APP_AUTH_DOMAIN,
   databaseURL: process.env.REACT_APP_DATABASE_URL,
@@ -9,4 +9,4 @@ const firebaseConfig = {
   measurementId: process.env.REACT_APP_MESSAGING_SENDER_ID,
 };
 
-export default firebaseConfig;
+export default config;
