+++ src/Api/firebase/index.ts
@@ -1,6 +1,7 @@
 import firebase from 'firebase/app';
 import 'firebase/auth';
-import config from './firebase/config';
+
+import config from './config';
 
 firebase.initializeApp(config);
